+++ src/senaite/storage/browser/theme/icons.py
@@ -1,0 +1,50 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of SENAITE.STORAGE.
+#
+# SENAITE.STORAGE is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by the Free
+# Software Foundation, version 2.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+# details.
+#
+# You should have received a copy of the GNU General Public License along with
+# this program; if not, write to the Free Software Foundation, Inc., 51
+# Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+#
+# Copyright 2019-2020 by it's authors.
+# Some rights reserved, see README and LICENSE.
+
+import os
+
+from plone.resource.interfaces import IResourceDirectory
+from senaite.core.browser.globals.interfaces import IIconProvider
+from senaite.core.browser.globals.interfaces import ISenaiteTheme
+from zope.component import adapts
+from zope.component import getUtility
+from zope.interface import implementer
+
+ICON_BASE_URL = "++plone++senaite.storage.static/assets/icons"
+
+
+@implementer(IIconProvider)
+class IconProvider(object):
+    adapts(ISenaiteTheme)
+
+    def __init__(self, view, context):
+        self.view = view
+        self.context = context
+
+    def icons(self):
+        icons = {}
+        static_dir = getUtility(
+            IResourceDirectory, name=u"++plone++senaite.storage.static")
+        icon_dir = static_dir["assets"]["icons"]
+        for icon in icon_dir.listDirectory():
+            name, ext = os.path.splitext(icon)
+            icons[name] = "{}/{}".format(ICON_BASE_URL, icon)
+            icons[icon] = "{}/{}".format(ICON_BASE_URL, icon)
+        return icons